requirements.txt
requests
python-dotenv
streamlit
pandas
matplotlib